+++ noodles-bcf/src/io/compression_method.rs
@@ -1,0 +1,8 @@
+/// A BCF compression method.
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+pub enum CompressionMethod {
+    /// No compression.
+    None,
+    /// BGZF.
+    Bgzf,
+}