+++ noodles-bcf/src/reader/builder.rs
@@ -7,22 +7,38 @@ use std::{
 use noodles_bgzf as bgzf;
 
 use super::Reader;
+use crate::io::CompressionMethod;
 
 /// A BCF reader builder.
 #[derive(Default)]
-pub struct Builder;
+pub struct Builder {
+    compression_method: Option<CompressionMethod>,
+}
 
 impl Builder {
+    /// Sets the compression method.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use noodles_bcf::{io::CompressionMethod, reader::Builder};
+    /// let builder = Builder::default().set_compression_method(CompressionMethod::Bgzf);
+    /// ```
+    pub fn set_compression_method(mut self, compression_method: CompressionMethod) -> Self {
+        self.compression_method = Some(compression_method);
+        self
+    }
+
     /// Builds a BCF reader from a path.
     ///
     /// # Examples
     ///
     /// ```no_run
     /// use noodles_bcf::reader::Builder;
-    /// let reader = Builder.build_from_path("sample.bcf")?;
+    /// let reader = Builder::default().build_from_path("sample.bcf")?;
     /// # Ok::<_, std::io::Error>(())
     /// ````
-    pub fn build_from_path<P>(self, src: P) -> io::Result<Reader<bgzf::Reader<File>>>
+    pub fn build_from_path<P>(self, src: P) -> io::Result<Reader<Box<dyn Read>>>
     where
         P: AsRef<Path>,
     {
@@ -39,10 +55,15 @@ impl Builder {
     /// use noodles_bcf::reader::Builder;
     /// let reader = Builder::default().build_from_reader(io::empty());
     /// ```
-    pub fn build_from_reader<R>(self, reader: R) -> io::Result<Reader<bgzf::Reader<R>>>
+    pub fn build_from_reader<'r, R>(self, reader: R) -> io::Result<Reader<Box<dyn Read + 'r>>>
     where
-        R: Read,
+        R: Read + 'r,
     {
-        Ok(Reader::new(reader))
+        let inner: Box<dyn Read> = match self.compression_method {
+            Some(CompressionMethod::Bgzf) | None => Box::new(bgzf::Reader::new(reader)),
+            Some(CompressionMethod::None) => Box::new(reader),
+        };
+
+        Ok(Reader::from(inner))
     }
 }